[DOP-8140] Split base DBConnection classes to submodules
