+++ onetl/connection/db_connection/kafka/dialect.py
@@ -18,7 +18,8 @@ from __future__ import annotations
 import logging
 
 from onetl._util.spark import get_spark_version
-from onetl.connection.db_connection.db_connection import BaseDBConnection, DBConnection
+from onetl.base import BaseDBConnection
+from onetl.connection.db_connection.db_connection.dialect import DBDialect
 from onetl.connection.db_connection.dialect_mixins import (
     SupportColumnsNone,
     SupportDfSchemaNone,
@@ -38,7 +39,7 @@ class KafkaDialect(  # noqa: WPS215
     SupportWhereNone,
     SupportTableWithoutDBSchema,
     SupportHWMExpressionNone,
-    DBConnection.Dialect,
+    DBDialect,
 ):
     valid_hwm_columns = {"offset", "timestamp"}
